+++ js_modules/dagit/packages/eslint-config/CHANGES.md
@@ -1,3 +1,8 @@
+## 1.0.2 (April 13, 2022)
+
+- Enable `object-shorthand`
+- Add lint and prettier to this package
+
 ## 1.0.1 (April 7, 2022)
 
 - Fix dependencies